+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.INSOPTAndroidPractice"
         tools:targetApi="31">
         <activity
-            android:name=".HomeActivity"
+            android:name=".presentation.home.HomeActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.signup.SignUpActivity"
