+++ app/src/main/java/org/sopt/sample/presentation/home/HomeFragment.kt
@@ -1,10 +1,11 @@
-package org.sopt.sample
+package org.sopt.sample.presentation.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import org.sopt.sample.R
 import org.sopt.sample.adapter.RepoAdapter
 import org.sopt.sample.data.Repo
 import org.sopt.sample.databinding.FragmentHomeBinding
