+++ app/src/main/java/org/sopt/sample/presentation/home/SearchFragment.kt
@@ -1,4 +1,4 @@
-package org.sopt.sample
+package org.sopt.sample.presentation.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
