+++ app/src/main/res/layout/activity_login.xml
@@ -3,6 +3,11 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
+    <data>
+        <variable
+            name="viewModel"
+            type="org.sopt.sample.presentation.login.LoginViewModel" />
+    </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_height="match_parent"
@@ -75,9 +80,9 @@
             app:layout_constraintTop_toBottomOf="@+id/btn_login" />
 
         <androidx.constraintlayout.helper.widget.Flow
+            android:layout_width="0dp"
             android:layout_height="0dp"
             android:layout_marginTop="40dp"
-            android:layout_width="0dp"
             android:orientation="horizontal"
             app:constraint_referenced_ids="btn_login, btn_signup"
             app:flow_horizontalGap="12dp"
