+++ app/src/main/res/layout/fragment_gallery.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GalleryFragment">
+    tools:context=".presentation.home.GalleryFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_gallery"
